% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{pcapred.largedata-package}
\alias{pcapred.largedata}
\alias{pcapred.largedata-package}
\title{pcapred.largedata: Data for pcapred}
\description{
Use with package "pcapred", available at https://github.com/danjlawson/pcapred.
}
\details{
This is the optional large data reference data for "pcapred", written by Daniel Lawson (dan.lawson@bristol.ac.uk) and supporting a paper by Aliya Sarmanova, Tim Morris and Daniel Lawson (to appear). It contains reference data for the UK Biobank PCs with up to 200 PCs in \code{\link{ukb_pcs_200}} and 100 PCs in \code{\link{ukb_pcs_100}}.

See \url{https://github.com/danjlawson/pcapred} for details.
}
\examples{
\dontrun{
library("pcapred")
ref=readreference(pcapred.largedata::ukb_pcs_200())
}


}
\references{
To follow
}
\seealso{
\code{\link[pcapred]{readreference}}
}
\author{
\strong{Maintainer}: Daniel Lawson \email{dan.lawson@bristol.ac.uk}

}
\keyword{internal}
