% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ukb_pcs_100}
\alias{ukb_pcs_100}
\title{UK Biobank reference data for 100 PCs}
\usage{
ukb_pcs_100()
}
\value{
A character string to be passed to \code{\link[pcapred]{readreference}} as the file name.
}
\description{
The extdata directory contains three files with the file root \code{ukb_pca_100} that are read together with the \code{\link[pcapred]{readreference}} command. This function gets their location on your system.
}
\keyword{datasets}
